fix some testcases & refactor
