+++ NestedText/Api/NestedTextDeserializeException.cs
@@ -1,5 +1,9 @@
-﻿namespace NestedText;
+﻿using NestedText.Cst;
 
+namespace NestedText;
+
+
+public record ParsingError(int LineNumber, int ColumnNumber, string Message);
 public class NestedTextDeserializeException : Exception
 {
     public NestedTextDeserializeException()
