+++ NestedText/Api/NestedTextSerializeException.cs
